Added start of double battle vertical slice test
- Added check to prevent fainted or inactive Pokemon from using their
move
- Changed how the branch probability limit worked so events that would
cause one branch to be above the limit and the other below no longer
split
- Moved the vertical slice tests to their own file
- Simplified how the ideal number of outcomes is calculated
- Fixed incorrect calculations for whether an outcome would be included
in a given test configuration
  - It only worked before because the old wrong number and what's
correct coincidentally led the variables to be same value
